fix a bug which make a common 404 page
The app crashes whenever a user tries to visit any newly added
(or non-existent) poll details page directly via the address bar.

Remember that the newly added poll would be non-existent at their URL
due to the nature of how the backend is set up in the application.
The value of questions[question_id] would be undefined for such polls.